+++ TGEngine/game_content/Mesh.cpp
@@ -7,7 +7,7 @@ void Mesh::consume(VertexBuffer * vrt, IndexBuffer * ind) {
 		this->offsets[i].offset += this->first_index;
 		render_offset.push_back(this->offsets[i]);
 	}
-	uint32_t vertex_offset = vrt->count_of_points;
+	uint32_t vertex_offset = (uint32_t)vrt->count_of_points;
 	vrt->addAll(this->vertices.data(), this->vertices.size());
 	for(uint32_t nt : this->indices)
 	{
